refactor: change data type

+++ headers/threaded/PrimePathThreaded.h
@@ -6,7 +6,7 @@ class PrimePathThreaded: public IPathBuilder {
 public:
     void displayPath(const std::string &start, const std::string &end, Graph* graph);
 private:
-    static bool isPrime(int n);
+    static bool isPrime(size_t n);
     struct State {
         std::string node;
         size_t weight;
